pyproject.toml
[project]
name = "nhm-llm"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10,<3.14"
dependencies = [
    "huggingface-hub>=1.8.0",
    "llama-cpp-python>=0.3.19",
]
